requirements.txt
# local package
-e .

# external package
click==8.1.7
langchain==0.0.344
openai==0.27.10
python-dotenv==1.0.0
setuptools==68.0.0
streamlit==1.29.0
streamlit-folium==0.16.0
streamlit_feedback==0.1.3
streamlit-option-menu==0.3.6
requests>=2.31.0
pydantic==2.3.0
watchdog==3.0.0
requests>=2.31.0